requirements.txt
beautifulsoup4>=4.13.4
blinker>=1.9.0
certifi>=2025.7.9
charset-normalizer>=3.4.2
click>=8.2.1
dnspython>=2.7.0
Flask>=3.1.1
flask-cors>=6.0.1
idna>=3.10
itsdangerous>=2.2.0
Jinja2>=3.1.6
jsonify>=0.5
lxml>=6.0.0
MarkupSafe>=3.0.2
packaging>=24.2
pyparsing>=3.1.4
requests>=2.32.4
setuptools>=70.3.0
soupsieve>=2.7
typing_extensions>=4.14.1
urllib3>=2.5.0
Werkzeug>=3.1.3
htmlmin>=0.1.12
python-whois>=0.8.0
